backend/requirements.txt
# Core Django and DRF
Django==5.1
djangorestframework
djangorestframework-simplejwt

# PostgreSQL Database Support
dj-database-url
psycopg2-binary

# Environment Variables Handling
python-dotenv

# Gunicorn for Running Django in Production Mode
gunicorn

# Optional: CORS Handling (if your React Native app makes API calls)
#django-cors-headers

# Optional: Security Enhancements
django-environ

# Optional: Debug Toolbar for Development
django-debug-toolbar